🎨 Fix conftest
